Safari-611.1.10/LayoutTests/fast/forms/state-restore-hidden-expected.txt

Bug 77391 - Hidden form elements do not save their state prior to form submission

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.f.hidden.value is "after"
PASS successfullyParsed is true

TEST COMPLETE
